Address Actual Durations Far Exceeding Threshold Durations (#7)
* Fix for issue 6

This changes the behavior of the package back to what it was originally intended to be which is that the intervals given as the Durations represents the time depth into the connection.  1min should trigger at 1 minute into the connection.  5min should trigger at 5min into the connection.

This does imply that timestamps need to be sequentially greater in the provided Durations.